myproject/requirements.txt
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
Pillow==8.3.2
